Correct trigger deactivation
Triggers didnt deactivate when Act_DeactTrig, and they did not reset their conditions after Deactivation.

+++ FLHook/2.0/plugins/flsr_plugin/missions/NpcArch.h
@@ -27,7 +27,7 @@ namespace Missions
 		Matrix orientation;
 		uint pilotJobId = 0;
 		uint startingObjId = 0;
-		float hitpoints = -1.0f;
+		int hitpoints = -1;
 		std::unordered_set<uint> labels;
 	};
 	typedef std::shared_ptr<MsnNpcArchetype> MsnNpcArchetypePtr;
